+++ package.json
@@ -14,7 +14,9 @@
   "license": "ISC",
   "dependencies": {
     "@types/express": "^4.17.3",
+    "@types/node-fetch": "^2.5.5",
     "@types/node-telegram-bot-api": "^0.40.2",
+    "axios": "^0.19.2",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "node-telegram-bot-api": "^0.40.0"
